erlang-r16 R16B01
Adds `--with-wxmac` to enable wxWidgets support and `--with-fop` to
build PDF documentation.

It also includes a fix for the manpages from the official erlang
formula two small fixes for Mac OS X 10.9 which no longer includes
iodbc in it's default sdk and only provides headers inside Xcode.app,
so we abuse configure's internal crosscompile vars to fix detection.

Closes #150.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>