Redact old Python versions.
Python support in Homebrew is tricky to get right, and the way the
most-recent versions of Python 2.7.x and 3.x are handled in core
change over time.

These older versions of Python have not kept up-to-date with these
layout changes, and have installation problems as-is.

In addition, these versions may not be getting important security
updates any more.

Removing these from the versions repo; if they are still needed
by any users they should be handled in a (different) external tap.

Closes #84.
Closes #62.
Closes #229.
Closes #167.
Closes https://github.com/Homebrew/homebrew/issues/27208