subversion16: modernise
The deps url is dead, and the universal build options are no
longer recognised by Homebrew.

Tried to replace the former with
independent vendoring of each resource as per dictated by the
get-deps script in this subversion version, and including Homebrew
deps for the other elements.

Replaced the latter with modernised language.

Closes #713.

Signed-off-by: Dominyk Tiller <dominyktiller@gmail.com>