hdf4: stop supporting fortran
The Fortran element of this formula hasn't been installing as expected for a long
time, and nobody knows why or cares enough to want to spend not an insignificant
amount of hours researching it and fixing it, if it can be fixed.

If you know why it isn't working as intended, please submit a PR with appropriate
fixes. Please don't just file issues asking for someone else to re-add that support.

Thanks!

Closes #683