node4-lts: updated head branch
The v4.x branch would always point to the sources of the latest stable release
of the v4.x release line making it redundant with stable.

The v4.x-staging branch holds commits planned for the next v4.x release before
the actual release better matching the behavior of a 'head' branch.

Closes #1060.

Signed-off-by: Mike McQuaid <mike@mikemcquaid.com>