openssl098: remove
As we've been warning for months now, OpenSSL 0.9.8 was officially deprecated
upstream at the end of December 2015.

The OpenSSL team stated earlier and have since reiterated that due to this they
aren't even doing vulnerability analysis on the 0.9.8 branch, so we are blind in
terms of knowing the impact of vuln announcements going forwards.

Today OpenSSL announced there will be new OpenSSL releases for the 1.0.1 and
1.0.2 branches on Thursday to resolve two security issues.

Taking that on board, and the official deprecation deadline passing, this formula
is being removed from the Homebrew organisation. You are more than welcome to
take it and support it externally, but at this point the 0.9.8 branch is supported
by pretty much nobody and _is_ going to gradually become increasingly risky.

Please strongly consider moving to at least the 1.0.1 branch of OpenSSL in your
projects. We support a formula for that in this very tap.